requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
motor>=3.1.2
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6
email-validator>=2.0.0
pymongo>=4.3.3
python-dotenv>=1.0.0
httpx>=0.24.0
authlib>=1.2.1
requests>=2.31.0
cloudinary>=1.36.0
